+++ Source/Enum/Scopes.ps1
@@ -1,4 +1,4 @@
-enum ConnectionScopes {
+enum Scopes {
     # Scopes in which a connection can be invoked
 
     Undefined