+++ Source/Public/New-MyRMInventory.ps1
@@ -81,6 +81,7 @@ function New-MyRMInventory {
                             "C:\Windows\System32\OpenSSH\ssh.exe",
                             "-l <user> -p <port> <host>",
                             22,
+                            [Scopes]::Console,
                             "OpenSSH (Microsoft Windows feature)"
                         )
                     )
@@ -91,6 +92,7 @@ function New-MyRMInventory {
                             "putty.exe",
                             "-ssh -P <port> <user>@<host>",
                             22,
+                            [Scopes]::External,
                             "PuTTY using SSH protocol"
                         )
                     )
@@ -101,6 +103,7 @@ function New-MyRMInventory {
                             "C:\Windows\System32\mstsc.exe",
                             "/v:<host>:<port> /fullscreen",
                             3389,
+                            [Scopes]::External,
                             "Microsoft Remote Desktop"
                         )
                     )