+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to MyRemoteManager is documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [develop]
+
+### Added
+
+- Scopes in which connections can be invoked
+- `DefaultScope` parameter in `Add-MyRMClient` to define a default scope associated to the new client
+
 ## [0.2.0]
 
 ### Added